Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> How to repro? It no longer crashs on my device after moving to class main

on x2, I would randomly have race conditions with "LiveDisplay failed to get configuration" in logcat and visually after lockscreen I would see "phone is starting" and if the screen turned off and if it turned on afterwards it would fail to render.

I'll be testing this patch out, thanks Paul.

Patch-set: 2
CC: Gerrit User 19252 <19252@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>