Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-2
> > 
> > This needlessly complicates the code, and I haven't seen any reason for this to be done other than "It's c++"
> 
> This makes the code easier to read IMO. Use of RAII also reduces the likelihood of resource leakage, as objects clean up their resources as they are destroyed.

Personally, I find it harder to read because it obscures the actual functional bits of the code. I think the SDMController code is pretty gross to read, but that's not a blocker. I also think that we're defining a lot of unnecessary operators for sdm_disp_mode that we don't even use. I'll look at that more closely later, though because on top of that there are two issues:

1. This has been proposed as a fix for the LD hal sometimes not starting correctly and causing a hang on bootani. The HAL not init'ing can and should be fixed separately, switching to RAII just obscures the actual underlying issue.

2. It's also been established that this breaks the existing expectation that PictureAdjustment can be used without DisplayModes. The breakage that we fix in a later patch (278932) should be addressed in this one, without a hack like what is currently proposed (i.e. the design is flawed if we're introducing a dependency between those two that we have to fix in such a kludgy manner).

Patch-set: 3
Reviewer: Gerrit User 11801 <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>