Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Nothing in that page says we _need_ to define them, and that page is pretty clear that not all circumstances require defining all 3/5. We really only need the normal constructor and destructor, because they are nothing but intermediate types.
> 
> These words again reflect that you know _nothing_ about C++ memory management, possibly any system-level language w/o GC, and you're even too lazy to google related blogs or even StackOverflow posts when someone else already pointed out the topic. I'll stop the meaningless debate for now because you are so obstinate and nearly always refuse to take my advice.
> 

I'm used to dealing with managing memory manually and not using C++ constructs to do it (see: all of the existing LD hal), and I''ll admit I'm a bit stuck in my ways. However, resorting to insults/snark without explaining is really not appropriate here and doesn't help you at all. The thing to keep in mind when proposing changes is that it's up to *you* to justify your change, not anyone else, and you should be able to justify it respectfully.

Anyways, onto technical things:

Here's what I see: creating the vector calls the constructor, and then we pass the internal array (which is a pointer, so I don't see why any of the copy constructors would be invoked) to the sdm backend (which knows absolutely nothing about our type, so it just writes it directly). Then the objects are destroyed by calling the deconstructor when the vector goes out of scope. We need only the deconstructor and the constructor afaict, and that's exactly how we did it in livedisplay@1.0. The article itself says "If a class requires a user-defined destructor, a user-defined copy constructor, or a user-defined copy assignment operator, it *almost* certainly requires all three" - we only use the construction and destruction here to simplify having to use new/free, which is fine, but defining the copy constructors and copy assignments is unnecessary imo for a type which is not even meant to be exposed. If I'm misunderstanding something, please let me know where _without_ resorting to insults.

> > The program already checks if the backend is not up and restarts the service if it is not
> 
> It's _and_ but should be _or_.

That doesn't make any sense, though. PictureAdjustment should be always supported, and DisplayModes is sometimes supported. You're saying we should fail if both features are not reported as supported. Why would we do that?

Patch-set: 3
Reviewer: Gerrit User 11801 <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>